requirements.txt
certifi==2019.6.16
chardet==3.0.4
configparser==3.7.4
dj-database-url==0.5.0
Django==2.2
djangorestframework==3.9.4
future==0.17.1
gunicorn==19.9.0
idna==2.8
oauthlib==3.0.1
psycopg2==2.8.3
pycryptodomex==3.4.7
PyJWT==1.6.4
python-dateutil==2.8.0
python-dotenv==0.10.3
python-twitter==3.5
pytz==2018.9
requests==2.22.0
requests-oauthlib==1.2.0
sqlparse==0.3.0
urllib3==1.25.3
whitenoise==4.1.2
